remove old llama monkey patch + rename mistral_monkey_patch
